Differentiation for static files
Checks the URI to see if it specifies an existing file, if so then takes the
same actions that it did in the previous homework.  Otherwise it checks the
list of Plugins to see if there is a Plugin that matches the first part of the
URI.  Then it calls the route request method for that plugin and then returns
its response which is handled by the server as in the tutorial.  If the URI is
neither a static file or a plugin then it just return a 404.